Threshold/requirements.txt
# Python requirements for threshold rules processor
psycopg2-binary==2.9.9
python-dotenv==1.0.0
